requirements.txt
streamlit==1.44.1
llama-index==0.12.29
python-dotenv==1.1.0
asyncio==3.4.3